Extract runner debug helpers
